s3 uploads work; gql blocked
